backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
google-generativeai==0.3.2
pytesseract==0.3.10
Pillow==10.1.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
aiofiles==23.2.1
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
textblob==0.17.1
nltk==3.8.1
redis==5.0.1
fakeredis==2.20.0
